Update patch set 3
Patch Set 3:

So what are we going to do? 
I added the pn544 addition for find7op and if this change is needed for other devices to be able to use pn544 i am fine with it. With the change in the kernel it will build and work for find7op.

Patch-set: 3